requirements.txt
./byepy-0.1.0-py2.py3-none-any.whl
psycopg2-binary==2.9.1
python-dotenv==0.19.2
types-psycopg2==2.9.0
